♻️  refactor(astral): refactor code to use persistent variable to store colorscheme
Plugin now uses vim.g.COLORTHEME to store the theme rather than storing it in a file. Performance improvement!

Removed restore_colors option as it is redundant with that being the plugin's only purpose
