+++ README.md
@@ -39,7 +39,6 @@ The default values for `astral` configuration are:
 
 ```lua
 local config = {
-    restore_colors = true, -- Enable or disable automatic colorscheme restoration
     fallback_themes = { -- Default fallback themes
         "catppuccin",
         "tokyonight",
@@ -59,7 +58,6 @@ require("lazy").setup({
         "astral/astral.nvim",
         version = "*", -- Pin to GitHub releases
         opts = {
-            restore_colors = true, -- Set to false to disable colorscheme restoration
             fallback_themes = { -- Customize fallback themes
                 "catppuccin",
                 "tokyonight",
@@ -81,17 +79,16 @@ require("lazy").setup({
 
 ## Troubleshooting
 
-If you encounter issues with `astral.nvim`, here are some common problems and solutions:
+If you encounter issues with `astral.nvim`, here are some common problems and
+solutions:
 
 1. **Colorscheme not restoring on startup**:
    - Ensure `restore_colors` is set to `true` in the configuration.
-   - Check if the `settings.json` file exists in the `data` path
-   (`vim.fn.stdpath("data") .. "/lazy/astral.nvim/settings.json"`).
-   If not, the plugin should create it automatically.
+   - Check if `vim.g.COLORTHEME` is set and holds the correct colorscheme name.
 
 2. **Fallback themes not applied**:
    - Verify that the `fallback_themes` array in the configuration contains
-   valid colorscheme names.
+     valid colorscheme names.
    - Ensure that the fallback themes are installed and available in Neovim.
 
 3. **Unknown Astral command error**:
@@ -100,14 +97,14 @@ If you encounter issues with `astral.nvim`, here are some common problems and so
      - `:Astral restore`
 
 4. **Colorscheme changes not saved**:
-   - Ensure there are no errors in the autocommand that triggers
-   on the `ColorScheme` event.
-   - Check if the `settings.json` file is writable and that your
-   Neovim configuration allows writing to this path.
+   - Ensure there are no errors in the autocommand that triggers on the
+     `ColorScheme` event.
+   - Check if `vim.g.COLORTHEME` is being updated correctly when you change the
+     colorscheme.
 
 If the issue persists, please open an issue on the
-[GitHub repository](https://github.com/rootiest/astral.nvim) with details
-about your configuration and the problem you are facing.
+[GitHub repository](https://github.com/rootiest/astral.nvim)
+with details about your configuration and the problem you are facing.
 
 ## License
